extract sendProxyRequest to module (#185)
